Populate StatePane with detectable factories
